+++ lib/runners/processor/metrics_fileinfo.rb
@@ -20,6 +20,11 @@ module Runners
     end
 
     def analyze(changes)
+      # Generate pre-computed cache for Git metadata access (https://git-scm.com/docs/git-commit-graph)
+      # This improves the performance of access to Git metadata for a large repository.
+      # You can see the efficacy here: https://github.com/sider/runners/issues/2028#issuecomment-776534408
+      capture3!("git", "commit-graph", "write", "--reachable", "--changed-paths", "--no-progress")
+
       target_files = changes.changed_paths.map(&:to_path)
       analyze_last_committed_at(target_files)
       analyze_lines_of_code(target_files)
@@ -58,11 +63,11 @@ module Runners
       text_files = targets.select { |f| text_file?(f) }
       text_files.each_slice(1000) do |files|
         stdout, _ = capture3!("wc", "-l", *files)
-        lines = stdout.lines(chomp: true).tap do |l|
-          # wc command outputs total count when we pass multiple targets. remove it if exist
-          last_line = (l.last or break)
-          l.pop if last_line.match?(/^\d+ total$/)
-        end
+        lines = stdout.lines(chomp: true)
+
+        # `wc` command outputs total count when we pass multiple targets. remove it if exist
+        lines.pop if lines.last&.match?(/^\d+ total$/)
+
         lines.each do |line|
           fields = line.split(" ")
           loc = (fields[0] or raise)
@@ -78,7 +83,7 @@ module Runners
 
     def analyze_last_committed_at(targets)
       targets.each do |target|
-        stdout, _ = capture3!("git", "log", "-1", "--format=format:%aI", target)
+        stdout, _ = capture3!("git", "log", "-1", "--format=format:%aI", target, trace_stdout: false)
         last_committed_at[target] = stdout
       end
     end