+++ lib/runners/processor/shellcheck.rb
@@ -97,7 +97,8 @@ module Runners
       targets = Array(config_linter[:target] || DEFAULT_TARGET)
 
       # Via glob
-      globs = _ = targets.select { |glob| glob.is_a? String } # TODO: Ignored Steep error
+      # @type var globs: Array[String]
+      globs = targets.filter { |glob| glob.is_a? String }
       files_via_glob = Dir.glob(globs, File::FNM_EXTGLOB, base: current_dir.to_path).filter { |path| File.file?(path) }
 
       # Via shebang
@@ -140,7 +141,7 @@ module Runners
             end_column: comment[:endColumn],
           ),
           message: comment[:message],
-          links: ["https://github.com/koalaman/shellcheck/wiki/#{id}"],
+          links: ["#{analyzer_github}/wiki/#{id}"],
           object: {
             code: comment[:code],
             severity: comment[:level],