+++ test/analyzers_test.rb
@@ -3,8 +3,37 @@ require "test_helper"
 class AnalyzersTest < Minitest::Test
   include TestHelper
 
-  def analyzers
-    @analyzers ||= Runners::Analyzers.new
+  def test_each
+    id, analyzer = analyzers.each.next
+    assert_equal :brakeman, id
+    assert_equal "Brakeman", analyzer[:name]
+  end
+
+  def test_each_block
+    called = false
+
+    analyzers.each do |id, analyzer|
+      assert_equal :brakeman, id
+      assert_equal "Brakeman", analyzer[:name]
+      called = true
+      break
+    end
+
+    assert called
+  end
+
+  def test_map
+    names = analyzers.map { |id, analyzer| "#{id} => #{analyzer[:name]}" }
+    assert_equal "brakeman => Brakeman", names.first
+  end
+
+  def test_size
+    assert_instance_of Integer, analyzers.size
+  end
+
+  def test_include?
+    assert analyzers.include?(:eslint)
+    refute analyzers.include?("foo")
   end
 
   def test_name
@@ -34,8 +63,33 @@ class AnalyzersTest < Minitest::Test
     assert_raises(KeyError) { analyzers.doc(:foo) }
   end
 
+  def test_website
+    assert_equal "https://eslint.org", analyzers.website(:eslint)
+    assert_equal "https://eslint.org", analyzers.website("eslint")
+  end
+
+  def test_docker
+    assert_equal "https://hub.docker.com/r/sider/runner_eslint", analyzers.docker(:eslint)
+    assert_equal "https://hub.docker.com/r/sider/runner_eslint", analyzers.docker("eslint")
+  end
+
   def test_deprecated
+    assert analyzers.deprecated?(:tslint)
+    assert analyzers.deprecated?("tslint")
     refute analyzers.deprecated?(:eslint)
     refute analyzers.deprecated?("eslint")
+  end
+
+  def test_beta
+    assert analyzers.beta?(:pylint)
+    assert analyzers.beta?("pylint")
+    refute analyzers.beta?(:eslint)
+    refute analyzers.beta?("eslint")
+  end
+
+  private
+
+  def analyzers
+    @analyzers ||= Runners::Analyzers.new
   end
 end