+++ test/smokes/eslint/expectations.rb
@@ -1,6 +1,6 @@
 s = Runners::Testing::Smoke
 
-default_version = "7.18.0"
+default_version = "7.20.0"
 
 s.add_test(
   "no_config",
@@ -358,7 +358,7 @@ s.add_test(
   issues: [
     {
       message: <<~MESSAGE.strip,
-        Parsing error: Unexpected token, expected ";"
+        Parsing error: Missing semicolon
 
           1 | function bar() {
         > 2 |   var x = foo:
@@ -368,7 +368,7 @@ s.add_test(
           5 |   }
       MESSAGE
       links: [],
-      id: "d91b54561db04524f183f5f8dee752dcf1d4fcb0",
+      id: "a146e219901ecca16bd2c00873406719d6b503d5",
       path: "index.js",
       location: { start_line: 2, start_column: 14 },
       object: { severity: "error", category: nil, recommended: nil },
@@ -416,7 +416,7 @@ s.add_test(
       id: "@typescript-eslint/no-unused-vars",
       message: "'x' is assigned a value but never used.",
       links: %w[
-        https://github.com/typescript-eslint/typescript-eslint/blob/v4.14.1/packages/eslint-plugin/docs/rules/no-unused-vars.md
+        https://github.com/typescript-eslint/typescript-eslint/blob/v4.15.1/packages/eslint-plugin/docs/rules/no-unused-vars.md
       ],
       path: "index.ts",
       location: { start_line: 1, start_column: 7, end_line: 1, end_column: 8 },