+++ test/smokes/haml_lint/expectations.rb
@@ -16,6 +16,17 @@ s.add_test(
       git_blame_info: {
         commit: :_, line_hash: "2f024a2bdf291a1ab61e026140f7e709028266a8", original_line: 4, final_line: 4
       }
+    },
+    {
+      message: "Line is too long. [204/200]",
+      links: %W[https://github.com/sds/haml-lint/blob/v#{default_version}/lib/haml_lint/linter#linelength],
+      id: "LineLength",
+      path: "test.haml",
+      location: { start_line: 7 },
+      object: { severity: "warning" },
+      git_blame_info: {
+        commit: :_, line_hash: "416b409e6b7628da6f061cd2c159b03157c38192", original_line: 7, final_line: 7
+      }
     }
   ],
   analyzer: { name: "HAML-Lint", version: default_version }
@@ -60,8 +71,11 @@ s.add_test(
   issues: [
     {
       message: "Lint/UselessAssignment: Useless assignment to variable - `unused_variable`.",
-      links: %W[https://github.com/sds/haml-lint/blob/v#{default_version}/lib/haml_lint/linter#rubocop],
-      id: "RuboCop",
+      links: %W[
+        https://github.com/sds/haml-lint/blob/v#{default_version}/lib/haml_lint/linter#rubocop
+        https://docs.rubocop.org/rubocop/cops_lint.html#lintuselessassignment
+      ],
+      id: "RuboCop:Lint/UselessAssignment",
       path: "test.haml",
       location: { start_line: 3 },
       object: { severity: "warning" },
@@ -78,8 +92,7 @@ s.add_test(
   type: "success",
   issues: [
     {
-      message:
-        "Lint/UselessAssignment: Useless assignment to variable - `unused_variable`. (https://rubystyle.guide#underscore-unused-vars)",
+      message: "Useless assignment to variable - `unused_variable`. (https://rubystyle.guide#underscore-unused-vars)",
       links: %w[https://github.com/sds/haml-lint/blob/v0.34.0/lib/haml_lint/linter#rubocop],
       id: "RuboCop",
       path: "test.haml",
@@ -111,9 +124,12 @@ s.add_test(
     {
       path: "test.haml",
       location: { start_line: 5 },
-      id: "RuboCop",
+      id: "RuboCop:Performance/FlatMap",
       message: "Performance/FlatMap: Use `flat_map` instead of `map...flatten`.",
-      links: %W[https://github.com/sds/haml-lint/blob/v#{default_version}/lib/haml_lint/linter#rubocop],
+      links: %W[
+        https://github.com/sds/haml-lint/blob/v#{default_version}/lib/haml_lint/linter#rubocop
+        https://docs.rubocop.org/rubocop-performance/cops_performance.html#performanceflatmap
+      ],
       object: { severity: "warning" },
       git_blame_info: {
         commit: :_, line_hash: "fbeff8480acd0bfaa22199218b67510fdfa9c389", original_line: 5, final_line: 5
@@ -136,6 +152,17 @@ s.add_test(
       object: { severity: "warning" },
       git_blame_info: {
         commit: :_, line_hash: "da7cbdf991191af285867e39fd1239b6a030b465", original_line: 2, final_line: 2
+      }
+    },
+    {
+      message: "Line is too long. [84/80]",
+      links: %W[https://github.com/sds/haml-lint/blob/v#{default_version}/lib/haml_lint/linter#linelength],
+      id: "LineLength",
+      path: "hello.haml",
+      location: { start_line: 7 },
+      object: { severity: "warning" },
+      git_blame_info: {
+        commit: :_, line_hash: "53d100be96dfb32b768bb472f9f7da2409f73ab6", original_line: 7, final_line: 7
       }
     },
     {
@@ -176,7 +203,8 @@ s.add_test(
       }
     }
   ],
-  analyzer: { name: "HAML-Lint", version: "0.26.0" }
+  analyzer: { name: "HAML-Lint", version: "0.26.0" },
+  warnings: [{ message: "The option `linter.haml_lint.parallel` is ignored with HAML-Lint 0.26.0. Please update it to **0.36.0+** or use our default version.", file: "sider.yml" }]
 )
 
 s.add_test(